new file:   .dockerignore 	modified:   .gitignore 	modified:   Dockerfile 	new file:   README.md 	new file:   app/api/v1/endpoints/manga/__init__.py 	new file:   app/api/v1/endpoints/manga/senkuro.py 	new file:   app/api/v1/endpoints/ranobe/__init__.py 	new file:   app/api/v1/endpoints/ranobe/handlers/__init__.py 	new file:   app/api/v1/endpoints/ranobe/handlers/ranobe_me.py 	new file:   app/api/v1/endpoints/ranobe/handlers/ranobehub.py 	new file:   app/api/v1/endpoints/ranobe/handlers/ranobelib.py 	new file:   app/api/v1/endpoints/ranobe/handlers/ranobepoisk.py 	new file:   app/api/v1/endpoints/ranobe/handlers/ranobes.py 	new file:   app/api/v1/endpoints/ranobe/ranobe.py 	new file:   app/api/v1/endpoints/ranobe/utils/f2b.py 	modified:   app/main.py 	new file:   app/redis/redis.conf 	modified:   app/redis/redis_client.py 	new file:   app/templates/index.html 	new file:   data/certbot/conf/options-ssl-nginx.conf 	modified:   docker-compose.yml 	modified:   init-build.sh 	modified:   init-letsencrypt.sh 	renamed:    nginx-conf.d/proxy_params.conf -> nginx/conf.d/proxy_params.conf 	renamed:    nginx-conf.d/nginx_app.conf -> nginx/nginx.conf 	new file:   redis/data/dump.rdb 	modified:   requirements.txt

+++ init-build.sh
@@ -1,6 +1,31 @@
 #!/bin/bash
 
+# Логирование
+LOG_FILE="cleanup.log"
+exec > >(tee -a "$LOG_FILE") 2>&1
+
+echo "Начало очистки: $(date)"
+
+# Проверка наличия Docker
+if ! command -v docker &> /dev/null; then
+    echo "Docker не установлен. Установите Docker и повторите попытку."
+    exit 1
+fi
+
+# Удаление данных certbot
+echo "Удаление данных certbot..."
 sudo rm -rf data/certbot
+
+# Остановка и удаление всех контейнеров
+echo "Остановка и удаление всех контейнеров..."
 docker rm -f $(docker ps -a -q)
+
+# Удаление всех Docker-образов
+echo "Удаление всех Docker-образов..."
 docker rmi $(docker images -q)
+
+# Инициализация Let's Encrypt
+echo "Запуск init-letsencrypt.sh..."
 ./init-letsencrypt.sh
+
+echo "Очистка завершена: $(date)"
